+++ components/style/data.rs
@@ -9,7 +9,6 @@ use selector_impl::PseudoElement;
 use std::collections::HashMap;
 use std::hash::BuildHasherDefault;
 use std::sync::Arc;
-use std::sync::atomic::AtomicIsize;
 
 pub struct PersistentStyleData {
     /// The results of CSS styling for this node.
@@ -18,9 +17,6 @@ pub struct PersistentStyleData {
     /// The results of CSS styling for each pseudo-element (if any).
     pub per_pseudo: HashMap<PseudoElement, Arc<ComputedValues>,
                             BuildHasherDefault<::fnv::FnvHasher>>,
-
-    /// Information needed during parallel traversals.
-    pub parallel: DomParallelInfo,
 }
 
 impl PersistentStyleData {
@@ -28,22 +24,6 @@ impl PersistentStyleData {
         PersistentStyleData {
             style: None,
             per_pseudo: HashMap::with_hasher(Default::default()),
-            parallel: DomParallelInfo::new(),
-        }
-    }
-}
-
-/// Information that we need stored in each DOM node.
-#[cfg_attr(feature = "servo", derive(HeapSizeOf))]
-pub struct DomParallelInfo {
-    /// The number of children that still need work done.
-    pub children_to_process: AtomicIsize,
-}
-
-impl DomParallelInfo {
-    pub fn new() -> DomParallelInfo {
-        DomParallelInfo {
-            children_to_process: AtomicIsize::new(0),
         }
     }
 }
